run/animation/agent/requirements.txt
google-auth>=2.23.0
google-cloud-storage>=2.18.0
langchain>=0.1.0
langchain-core>=0.1.27
langchain-google-vertexai>=2.0.0
langgraph>=0.0.31
vertexai>=1.43.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
gunicorn>=21.2.0
# Added for streaming support
aiohttp>=3.9.0
asyncio>=3.4.3
sse-starlette>=1.6.5